Refactored Converters in Seperate Namespace & Removed Unused Converter
