+++ Dukebox.Desktop/Converters/BoolToVisibilityConverter.cs
@@ -3,20 +3,20 @@ using System.Globalization;
 using System.Windows;
 using System.Windows.Data;
 
-namespace Dukebox.Desktop.Helper
+namespace Dukebox.Desktop.Converters
 {
-    public class BoolToVisibilityConverter : IValueConverter
+    public class ReverseBoolToVisibilityConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
             var boolValue = (bool)value;
-            return boolValue ? Visibility.Visible : Visibility.Hidden;
+            return boolValue ? Visibility.Hidden : Visibility.Visible;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
             var visibilityValue = (Visibility)value;
-            return visibilityValue == Visibility.Visible;
+            return visibilityValue != Visibility.Visible;
         }
     }
 }
