+++ Dukebox.Desktop/Converters/ReverseBoolToVisibilityConverter.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+
+namespace Dukebox.Desktop.Converters
+{
+    public class BoolToVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var boolValue = (bool)value;
+            return boolValue ? Visibility.Visible : Visibility.Hidden;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var visibilityValue = (Visibility)value;
+            return visibilityValue == Visibility.Visible;
+        }
+    }
+}
