+++ Dukebox.Desktop/Dukebox.Desktop.csproj
@@ -158,17 +158,16 @@
     <Compile Include="DukeboxUserSettings.cs" />
     <Compile Include="Factories\DukeboxUserSettingsFactory.cs" />
     <Compile Include="Helper\AudioCdHelper.cs" />
-    <Compile Include="Helper\ListDictionaryKeyToValueConverter.cs" />
-    <Compile Include="Helper\ReverseBoolToVisibilityConverter.cs" />
+    <Compile Include="Converters\ListDictionaryKeyToValueConverter.cs" />
+    <Compile Include="Converters\ReverseBoolToVisibilityConverter.cs" />
     <Compile Include="Interfaces\IAudioDriveListingViewModel.cs" />
     <Compile Include="Model\PreviewArtistOrAlbumMessage.cs" />
-    <Compile Include="Helper\BoolToVisibilityConverter.cs" />
-    <Compile Include="Helper\ArtistToCardPathConverter.cs" />
+    <Compile Include="Converters\BoolToVisibilityConverter.cs" />
+    <Compile Include="Converters\ArtistToCardPathConverter.cs" />
     <Compile Include="Helper\ImageResourceLoader.cs" />
     <Compile Include="Helper\ImageResources.cs" />
-    <Compile Include="Helper\ImageToImageSourceConverter.cs" />
     <Compile Include="Helper\ListSearchHelper.cs" />
-    <Compile Include="Helper\ShowTrackListingToHeightConverter.cs" />
+    <Compile Include="Converters\ShowTrackListingToHeightConverter.cs" />
     <Compile Include="Interfaces\IDukeboxUserSettings.cs" />
     <Compile Include="Interfaces\IArtistListingViewModel.cs" />
     <Compile Include="Interfaces\ILoadingScreenViewModel.cs" />
