--- Dukebox.Desktop/Helper/BoolToVisibilityConverter.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Globalization;
-using System.Windows;
-using System.Windows.Data;
-
-namespace Dukebox.Desktop.Helper
-{
-    public class ReverseBoolToVisibilityConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            var boolValue = (bool)value;
-            return boolValue ? Visibility.Hidden : Visibility.Visible;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            var visibilityValue = (Visibility)value;
-            return visibilityValue != Visibility.Visible;
-        }
-    }
-}
